Move log_crown_winners to utils/logging
Helper takes ``metagraph`` + ``block`` instead of the Validator instance,
matching ``miner_label`` / ``swap_label`` style. forward.py just imports
and calls.
